requirements.txt
django
django-extensions
dj-database-url
gunicorn
python-dotenv
hashids
djangorestframework
whitenoise
django-ratelimit
psycopg2==2.6.1
